Does not try to unlink file when it not possible
When cleaning up registries, there are some cases
where it's impossible to unlink the file:

1. The replicable record does not exist anymore;
2. The replicable file is stored on Object
   Storage, but the node is not configured to use
   Object Store;
3. Unrecognized replicable type;

+++ ee/app/services/geo/file_registry_removal_service.rb
@@ -6,9 +6,9 @@ module Geo
 
     LEASE_TIMEOUT = 8.hours.freeze
 
-    # It's possible that LfsObject or Ci::JobArtifact record does not exist anymore
-    # In this case, you need to pass file_path parameter explicitly
-    #
+    # There is a possibility that the replicable's record does not exist
+    # anymore. In this case, you need to pass the file_path parameter
+    # explicitly.
     def initialize(object_type, object_db_id, file_path = nil)
       @object_type = object_type.to_sym
       @object_db_id = object_db_id
@@ -26,9 +26,11 @@ module Geo
           break
         end
 
-        if File.exist?(file_path)
+        if file_path && File.exist?(file_path)
           log_info('Unlinking file', file_path: file_path)
           File.unlink(file_path)
+        else
+          log_error('Unable to unlink file because file path is unknown. A file may be orphaned', object_type: object_type, object_db_id: object_db_id)
         end
 
         log_info('Removing file registry', file_registry_id: file_registry.id)
@@ -61,6 +63,7 @@ module Geo
       strong_memoize(:file_path) do
         next @object_file_path if @object_file_path
         # When local storage is used, just rely on the existing methods
+        next if file_uploader.nil?
         next file_uploader.file.path if file_uploader.object_store == ObjectStorage::Store::LOCAL
 
         # For remote storage more juggling is needed to actually get the full path on disk
@@ -85,10 +88,19 @@ module Geo
         else
           raise NameError, "Unrecognized type: #{object_type}"
         end
+      rescue RuntimeError, NameError, ActiveRecord::RecordNotFound => err
+        # When cleaning up registries, there are some cases where
+        # it's impossible to unlink the file:
+        #
+        # 1. The replicable record does not exist anymore;
+        # 2. The replicable file is stored on Object Storage,
+        #    but the node is not configured to use Object Store;
+        # 3. Unrecognized replicable type;
+        #
+        log_error('Could not build uploader', err.message)
+
+        nil
       end
-    rescue NameError, ActiveRecord::RecordNotFound => err
-      log_error('Could not build uploader', err.message)
-      raise
     end
 
     def lease_key
